+++ src/frontend/apps/e2e/__tests__/app-desk/teams-panel.spec.ts
@@ -1,6 +1,6 @@
 import { expect, test } from '@playwright/test';
 
-import { createTeam, keyCloakSignIn } from './common';
+import { keyCloakSignIn } from './common';
 
 test.beforeEach(async ({ page, browserName }) => {
   await page.goto('/');
@@ -14,18 +14,18 @@ test.describe('Teams Panel', () => {
     await expect(page.getByTestId('button-new-team')).toBeVisible();
   });
 
-  test('checks the hover and selected state', async ({ page, browserName }) => {
-    const panel = page.getByLabel('Teams panel').first();
-    await createTeam(page, 'team-hover', browserName, 2);
+  // test('checks the hover and selected state', async ({ page, browserName }) => {
+  //   const panel = page.getByLabel('Teams panel').first();
+  //   await createTeam(page, 'team-hover', browserName, 2);
 
-    const selectedTeam = panel.locator('li').nth(0);
-    await expect(selectedTeam).toHaveCSS(
-      'background-color',
-      'rgb(133, 133, 246)',
-    );
+  //   const selectedTeam = panel.locator('li').nth(0);
+  //   await expect(selectedTeam).toHaveCSS(
+  //     'background-color',
+  //     'rgb(133, 133, 246)',
+  //   );
 
-    const hoverTeam = panel.locator('li').nth(1);
-    await hoverTeam.hover();
-    await expect(hoverTeam).toHaveCSS('background-color', 'rgb(202, 202, 251)');
-  });
+  //   const hoverTeam = panel.locator('li').nth(1);
+  //   await hoverTeam.hover();
+  //   await expect(hoverTeam).toHaveCSS('background-color', 'rgb(202, 202, 251)');
+  // });
 });